django/requirements.txt
django==2.1.7  # Web framework
richtypo==0.2.3  # Typograph text
pygments==2.3.1  # Highlight code
werkzeug==0.15.2  # Cool debugger
gunicorn==19.8.1  # WSGI HTTP server
django_ace==1.0.5  # Ace editor in admin site
django-filer==1.5.0  # Easy files in admin site
django-filter==2.1.0  # Filters for REST API
psycopg2-binary==2.7.4  # PostgreSQL driver
django-softhyphen==1.1.0  # Hyphenate HTML code
django-extensions==2.1.6  # Needed to run Werkzeug debugger
djangorestframework==3.9.2  # Generate REST API
django-constance[database]==2.2.0  # Dynamic settings in admin site
